% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cd2030.vaxx-package}
\alias{cd2030.vaxx}
\alias{cd2030.vaxx-package}
\title{cd2030.vaxx: The Countdown to 2030 Vaxx App}
\description{
The Countdown to 2030 Vaxx Shiny app: load and check a country's health facility data, choose denominators, analyse coverage, equity, mortality, service use and the health system, and build reports. Run it with run_app().
}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com} (\href{https://orcid.org/0009-0003-6159-1107}{ORCID})

Other contributors:
\itemize{
  \item African Population and Health Research Center (APHRC) [copyright holder, funder]
}

}
\keyword{internal}
